+++ server/schema/schema.js
@@ -246,6 +246,26 @@ const Mutation = new GraphQLObjectType({
             }
         },
 
+        //Remove User
+        RemoveUser: {
+            type: UserType,
+            args: {
+                id: {type: new GraphQLNonNull(GraphQLString)}
+
+            },
+            resolve(parent, args) {
+                let removedUser = User.findByIdAndRemove(
+                    args.id
+                ).exec();
+
+                if(!removedUser){
+                    throw new("Error");
+                }
+
+                return removedUser;
+            }
+        },
+
         //create post mutation
         CreatePost: {
             type: PostType,